Expand packet size support to 32-bit (int) length fields
Changed packet length properties from ushort to int, raising the maximum supported packet size from 65KB to over 2GB. Updated PacketConstants, FrameBase, and PacketBase accordingly. Adjusted IS_VALID_PACKET_SIZE and removed obsolete size validation in FragmentOptions. Improved SAEA_RECEIVE_EXACTLY_ASYNC for better performance. Updated documentation and performed minor formatting fixes.

+++ src/Nalix.Framework/DataFrames/FrameBase.cs
@@ -20,10 +20,10 @@ public abstract class FrameBase : IPacket
     /// <summary>
     /// Gets the total length of the serialized packet in bytes, including header and content.
     /// </summary>
-    [SerializeIgnore] public abstract ushort Length { get; }
+    [SerializeIgnore] public abstract int Length { get; }
 
     /// <inheritdoc/>
-    [SerializeIgnore] ushort IPacket.Length => this.Length;
+    [SerializeIgnore] int IPacket.Length => this.Length;
 
     /// <summary>
     /// Gets the magic number used to identify the packet format.
